+++ src/commands/should-not-have-class.js
@@ -1,0 +1,6 @@
+/**
+ * Ensure that the previous element does not have the given class.
+ */
+Cypress.Commands.add("shouldNotHaveClass", { prevSubject: "element" }, (subject, className) => {
+	return cy.wrap(subject).should("not.have.class", className);
+});